refactor pages separate from tests
