--- src/test/java/AddOwnerFormPage.java
@@ -1,4 +1,0 @@
-// Page URL: http://localhost:9966/petclinic/owners/new
-
-public class AddOwnerFormPage {
-}
