--- src/test/java/AddPetPage.java
@@ -1,4 +1,0 @@
-// Page URL: http://localhost:9966/petclinic/owners/2/pets/new.html
-
-public class AddPetPage {
-}
