--- src/test/java/EditOwnerPage.java
@@ -1,4 +1,0 @@
-// Page URL: http://localhost:9966/petclinic/owners/2/edit.html
-
-public class EditOwnerPage {
-}
