--- src/test/java/ErrorPage.java
@@ -1,4 +1,0 @@
-// Page URL: http://localhost:9966/petclinic/oups.html
-
-public class ErrorPage {
-}
