--- src/test/java/FindOwnersPage.java
@@ -1,4 +1,0 @@
-// Page URL: http://localhost:9966/petclinic/owners/find.html
-
-public class FindOwnersPage {
-}
