--- src/test/java/OwnersListingPage.java
@@ -1,4 +1,0 @@
-// Page URL: http://localhost:9966/petclinic/owners.html?lastName={search value}
-
-public class OwnersListingPage {
-}
