--- src/test/java/Utils.java
@@ -1,4 +1,0 @@
-public class Utils {
-    final static String BASE_URL = "localhost:9966/petclinic";
-    final static String CHROME_DRIVER_LOCATION = "chromedriver";
-}
