--- src/test/java/VetPage.java
@@ -1,4 +1,0 @@
-// Page URL: http://localhost:9966/petclinic/vets.html
-
-public class VetPage {
-}
