+++ src/test/java/webpages/AddOwnerFormPage.java
@@ -1,0 +1,4 @@
+package webpages;// Page URL: http://localhost:9966/petclinic/owners/new
+
+public class AddOwnerFormPage {
+}
