+++ src/test/java/webpages/AddPetPage.java
@@ -1,0 +1,4 @@
+package webpages;// Page URL: http://localhost:9966/petclinic/owners/2/pets/new.html
+
+public class AddPetPage {
+}
