+++ src/test/java/webpages/EditOwnerPage.java
@@ -1,0 +1,4 @@
+package webpages;// Page URL: http://localhost:9966/petclinic/owners/2/edit.html
+
+public class EditOwnerPage {
+}
