+++ src/test/java/webpages/ErrorPage.java
@@ -1,0 +1,4 @@
+package webpages;// Page URL: http://localhost:9966/petclinic/oups.html
+
+public class ErrorPage {
+}
