+++ src/test/java/webpages/NavBar.java
@@ -1,7 +1,8 @@
+package webpages;
+
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
-import org.openqa.selenium.support.PageFactory;
 
 public class NavBar extends PageObject{
     public NavBar(WebDriver driver) {
