+++ src/test/java/webpages/OwnerInfoPage.java
@@ -1,2 +1,4 @@
+package webpages;
+
 public class OwnerInfoPage {
 }
