+++ src/test/java/webpages/OwnersListingPage.java
@@ -1,0 +1,4 @@
+package webpages;// Page URL: http://localhost:9966/petclinic/owners.html?lastName={search value}
+
+public class OwnersListingPage {
+}
