+++ src/test/java/webpages/PageObject.java
@@ -1,3 +1,5 @@
+package webpages;
+
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.support.PageFactory;
