+++ src/test/java/webpages/Utils.java
@@ -1,0 +1,6 @@
+package webpages;
+
+public class Utils {
+    public final static String BASE_URL = "localhost:9966/petclinic";
+    public final static String CHROME_DRIVER_LOCATION = "chromedriver";
+}
