+++ src/test/java/webpages/VetPage.java
@@ -1,0 +1,4 @@
+package webpages;// Page URL: http://localhost:9966/petclinic/vets.html
+
+public class VetPage {
+}
